+++ griptape/schemas/base_schema.py
@@ -179,8 +179,11 @@ class BaseSchema(Schema):
             BaseVectorStoreDriver,
         )
         from griptape.events import EventListener
-        from griptape.memory.structure import Run
+        from griptape.memory import TaskMemory
+        from griptape.memory.structure import BaseConversationMemory, Run
+        from griptape.memory.task.storage import BaseArtifactStorage
         from griptape.structures import Structure
+        from griptape.tasks import BaseTask
         from griptape.tokenizers import BaseTokenizer
         from griptape.tools import BaseTool
         from griptape.utils import import_optional_dependency, is_dependency_installed
@@ -204,6 +207,9 @@ class BaseSchema(Schema):
                 "BaseMessageContent": BaseMessageContent,
                 "BaseDeltaMessageContent": BaseDeltaMessageContent,
                 "BaseTool": BaseTool,
+                "BaseTask": BaseTask,
+                "State": BaseTask.State,
+                "TaskMemory": TaskMemory,
                 "Usage": Message.Usage,
                 "Structure": Structure,
                 "BaseTokenizer": BaseTokenizer,
@@ -211,6 +217,8 @@ class BaseSchema(Schema):
                 "Reference": Reference,
                 "Run": Run,
                 "Sequence": Sequence,
+                "BaseConversationMemory": BaseConversationMemory,
+                "BaseArtifactStorage": BaseArtifactStorage,
                 # Third party modules
                 "Client": import_optional_dependency("cohere").Client if is_dependency_installed("cohere") else Any,
                 "GenerativeModel": import_optional_dependency("google.generativeai").GenerativeModel